# chore: Optimize table test performance