… packages/@react-spectrum/table/test/Table.test.js
… packages/@react-spectrum/table/test/TableTests.js
… packages/@react-spectrum/table/test/TreeGridTable.test.tsx