This optimizes the table tests to improve their performance. Before 91s, after 22s.

<img width="892" alt="image" src="https://github.com/user-attachments/assets/e1b98edc-d646-4978-9871-e3ba3d39f8e7" />

In circleci, test-19 went from 6m 33s on main to 3m 22s on this branch.

### Optimizations

* The version of JSDOM we use has a very slow implementation of `window.getComputedStyle` (https://github.com/jsdom/jsdom/issues/3234). This is supposedly fixed in newer versions but jest has not upgraded yet. In the meantime, we don't really need a full implementation, so I added a mock that just returns element.style without taking any other stylesheets into consideration. This was the biggest optimization.
* `getByLabelText` is also quite slow. This was used to find the select all checkbox. I added a `data-testid` instead which is much faster.
* Rendered fewer rows/columns. We set virtualizer to render a 1000x1000 rectangle, which contained hundreds of cells. Reducing this improves performance a lot. Also reduced the size of the collection itself in some cases.
* Moved the tests in `Table.test.js` into a separate file which can be reused with `TableNestedRows.test.js`. Because `Table.test.js` was imported into `TableNestedRows.test.js`, those tests actually ran twice before. Sorry for the large diff.